refactor(document-survey): rename util to utils
