+++ client/src/ui/molecules/document-survey/utils.ts
